+++ package.json
@@ -17,7 +17,10 @@
     "@prisma/client": "^5.3.1",
     "body-parser": "^1.20.2",
     "dotenv": "^16.3.1",
+    "ejs": "^3.1.9",
     "express": "^4.18.2",
-    "nodemon": "^3.0.1"
+    "express-session": "^1.17.3",
+    "nodemon": "^3.0.1",
+    "path": "^0.12.7"
   }
 }
