+++ public/styles/style.css
@@ -1,0 +1,45 @@
+body{
+    background-color: gray;
+}
+.content{
+    display: flex;
+    flex-direction:column;
+    align-items: center;
+}
+#form{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+#input-field{
+    border-radius: 3px;
+    border: none;
+    height: 30px;
+    width: 200px;
+    margin: 10px;
+
+    
+}
+#addButton{
+    display: inline-block;
+    border: none;
+    border-radius: 5px;
+    padding: 5px 40px;
+    background-color: white;
+    font-weight: bold;
+    text-decoration: none;
+    opacity: 0.7;
+
+}
+#addButton:hover{
+    background-color: rgb(215, 126, 126);
+    opacity: 1;
+}
+#head{
+}
+.content{
+
+}
+#fetchAll{
+    margin-bottom: 15px;
+}
