+++ views/todo/todo.ejs
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="styles/style.css">
+    <title>
+        <%=title%>
+    </title>
+
+</head>
+
+<body">
+    <div class="content">
+        <h1 id="head">TodoMatic V1</h1>
+        <form id="form" action="/todo" method="POST">
+            <div>
+                <input type="text" id="input-field" name="name" placeholder="name" required>
+                <input type="text" id="input-field" name="comment" placeholder="comment" required>
+            </div>
+            <button type="submit" id="addButton">Submit</button>
+        </form>
+        <ul id="fetchAll1">
+            <% items.forEach((item)=> { %>
+                <li id="fetchAll">
+                    <div>
+                        <%= item.name %>
+                    </div>
+                    <div>
+                        <%= item.comment %>
+                    </div>
+                    <div>
+                        <%= item.time %>
+                    </div>
+                </li>
+                <% }) %>
+        </ul>
+    </div>
+    </body>
+
+</html>
